+++ src/tree.h
@@ -16,9 +16,9 @@ extern int tree_get(const tree tptr, const TREE_KEY* key, unsigned long long* va
 extern int tree_set(const tree tptr, const TREE_KEY* key, unsigned long long value);
 extern int tree_del(tree tptr, const TREE_KEY* key);
 
-extern unsigned long tree_count(const tree tptr);
-extern int tree_stats(const tree tptr, unsigned long* trunks, unsigned long* branches, unsigned long* leafs);
-extern unsigned long tree_iter(const tree tptr, void* h, int (*f)(void* h, const TREE_KEY* key, unsigned long long* value));
+extern size_t tree_count(const tree tptr);
+extern int tree_stats(const tree tptr, size_t* trunks, size_t* branches, size_t* leafs);
+extern size_t tree_iter(const tree tptr, void* h, int (*f)(void* h, const TREE_KEY* key, unsigned long long* value));
 
 extern void tree_destroy(tree tptr);
